+++ src/Animate/Animate.uni.driver.d.ts
@@ -1,0 +1,5 @@
+import { BaseUniDriver } from 'wix-ui-test-utils/unidriver';
+
+export interface AnimateUniDriver extends BaseUniDriver {
+  getDelay(): Promise<string>;
+}