+++ src/MediaOverlay/test/MediaOverlay.visual.js
@@ -54,6 +54,21 @@ const tests = [
           ),
         },
       },
+      {
+        it: 'Without rounded borders',
+        props: {
+          media: (
+            <div
+              style={{
+                height: '100%',
+                background:
+                  'linear-gradient(to right top, #051937, #004d7a, #008793, #00bf72, #a8eb12)',
+              }}
+            />
+          ),
+          removeRoundedBorders: true,
+        },
+      },
     ],
   },
   {